Required change of Freshdesk Simple SSO url format
Just received a message from Freshdesk advising that the Simple SSO url format is changing to a form different than the one used by django-freshdesk.

"We noticed that you are using simple SSO and wanted to inform you that, moving forward, the format of your login details hash should be: Name, Shared secret key, Email address, and Timestamp (in that order). Please refer to this article and make the necessary changes ASAP.

As this is an important security update, we will stop supporting the old hashes by Thursday (28 April 2016) at 11 PM PDT.

We apologize in advance for the inconvenience. We are here to help you make these changes. If you need any assistance, you can just reply to this email or email support@freshdesk.com and we will be with you right away."